\name{makeRLearner}
\alias{makeRLearner}
\alias{makeRLearner.classif.constant}
\alias{makeRLearner.classif.dalda}
\alias{makeRLearner.classif.dalr}
\alias{makeRLearner.classif.damultinom}
\alias{makeRLearner.classif.dannet}
\alias{makeRLearner.classif.daqda}
\alias{makeRLearner.classif.dasvm}
\alias{makeRLearner.classif.FLXMCLconstant}
\alias{makeRLearner.classif.FLXMCLlda}
\alias{makeRLearner.classif.FLXMCLmultinom}
\alias{makeRLearner.classif.FLXMCLnnet}
\alias{makeRLearner.classif.FLXMCLqda}
\alias{makeRLearner.classif.FLXMCLsvm}
\alias{makeRLearner.classif.kda}
\alias{makeRLearner.classif.kmc}
\alias{makeRLearner.classif.mobConstantModel}
\alias{makeRLearner.classif.mobLdaModel}
\alias{makeRLearner.classif.mobMultinomModel}
\alias{makeRLearner.classif.mobNnetModel}
\alias{makeRLearner.classif.mobQdaModel}
\alias{makeRLearner.classif.mobSvmModel}
\alias{makeRLearner.classif.oslda}
\alias{makeRLearner.classif.osmultinom}
\alias{makeRLearner.classif.osnnet}
\alias{makeRLearner.classif.osqda}
\alias{makeRLearner.classif.ossvm}
\alias{makeRLearner.classif.wlda}
\alias{makeRLearner.classif.wqda}
\alias{makeRLearner.classif.wsvm}
\title{Methods for Creating \code{\link[mlr]{Learner}} Objects.}
\usage{
  \method{makeRLearner}{classif.constant} ()

  \method{makeRLearner}{classif.dalda} ()

  \method{makeRLearner}{classif.dalr} ()

  \method{makeRLearner}{classif.damultinom} ()

  \method{makeRLearner}{classif.dannet} ()

  \method{makeRLearner}{classif.daqda} ()

  \method{makeRLearner}{classif.dasvm} ()

  \method{makeRLearner}{classif.FLXMCLconstant} ()

  \method{makeRLearner}{classif.FLXMCLlda} ()

  \method{makeRLearner}{classif.FLXMCLmultinom} ()

  \method{makeRLearner}{classif.FLXMCLnnet} ()

  \method{makeRLearner}{classif.FLXMCLqda} ()

  \method{makeRLearner}{classif.FLXMCLsvm} ()

  \method{makeRLearner}{classif.kda} ()

  \method{makeRLearner}{classif.kmc} ()

  \method{makeRLearner}{classif.mobConstantModel} ()

  \method{makeRLearner}{classif.mobLdaModel} ()

  \method{makeRLearner}{classif.mobMultinomModel} ()

  \method{makeRLearner}{classif.mobNnetModel} ()

  \method{makeRLearner}{classif.mobQdaModel} ()

  \method{makeRLearner}{classif.mobSvmModel} ()

  \method{makeRLearner}{classif.oslda} ()

  \method{makeRLearner}{classif.osmultinom} ()

  \method{makeRLearner}{classif.osnnet} ()

  \method{makeRLearner}{classif.osqda} ()

  \method{makeRLearner}{classif.ossvm} ()

  \method{makeRLearner}{classif.wlda} ()

  \method{makeRLearner}{classif.wqda} ()

  \method{makeRLearner}{classif.wsvm} ()
}
\description{
  Methods for creating \code{\link[mlr]{Learner}} objects.
}

