\name{predictNode}
\alias{predictNode}
\title{Predict the ID of the Associated Terminal Node}
\usage{
  predictNode(.object, .newdata = NULL, ...)
}
\arguments{
  \item{.object}{Object of class \code{"mob"}.}

  \item{.newdata}{New data to predict.}

  \item{\dots}{Further arguments to
  \code{\link[party]{predict.mob}}.}
}
\value{
  A vector of IDs of terminal nodes.
}
\description{
  Predict the ID of the associated terminal node.
}

